Record when we’re using puma
Even if stats aren’t available.
